refactor: move config outside of api
