+++ scripts/getPositions.ts
@@ -31,12 +31,6 @@ const argv = yargs(process.argv.slice(2))
         return array.flatMap((v) => v.split(','))
       },
     },
-    getTokensInfoUrl: {
-      alias: 'g',
-      describe: 'URL to get token info from',
-      type: 'string',
-      default: 'https://api.mainnet.valora.xyz/getTokensInfo',
-    },
   })
   .parseSync()
 
@@ -55,12 +49,7 @@ function breakdownToken(token: Token): string {
 }
 
 void (async () => {
-  const positions = await getPositions(
-    argv.networkId,
-    argv.address,
-    argv.apps,
-    argv.getTokensInfoUrl,
-  )
+  const positions = await getPositions(argv.networkId, argv.address, argv.apps)
   console.log('positions', JSON.stringify(positions, null, ' '))
 
   console.table(
