+++ src/apps/beefy/positions.e2e.ts
@@ -1,17 +1,13 @@
 import hook from './positions'
 import { NetworkId } from '../../types/networkId'
-import { getConfig } from '../../api/config'
 import { getPositions } from '../../runtime/getPositions'
 
 describe('getPositionDefinitions', () => {
   it('should get the address definitions successfully', async () => {
-    const config = getConfig()
-
     const positions = await getPositions(
       NetworkId['op-mainnet'],
       '0x2b8441ef13333ffa955c9ea5ab5b3692da95260d',
       ['beefy'],
-      config.GET_TOKENS_INFO_URL,
     )
 
     expect(
