+++ src/apps/compound/positions.e2e.ts
@@ -1,17 +1,13 @@
-import { getConfig } from '../../api/config'
 import { getPositions } from '../../runtime/getPositions'
 import { NetworkId } from '../../types/networkId'
 import hook from './positions'
 
 describe('getPositionDefinitions', () => {
   it('should get the address definitions successfully for supply & collateral', async () => {
-    const config = getConfig()
-
     const positions = await getPositions(
       NetworkId['op-mainnet'],
       '0x2b8441ef13333ffa955c9ea5ab5b3692da95260d',
       ['compound'],
-      config.GET_TOKENS_INFO_URL,
     )
 
     const supplyPosition = positions.find((p) =>
@@ -26,13 +22,10 @@ describe('getPositionDefinitions', () => {
   })
 
   it('should get the address definitions successfully for debt & collateral', async () => {
-    const config = getConfig()
-
     const positions = await getPositions(
       NetworkId['arbitrum-one'],
       '0x2b8441ef13333ffa955c9ea5ab5b3692da95260d',
       ['compound'],
-      config.GET_TOKENS_INFO_URL,
     )
 
     const collateralPosition = positions.find((p) =>
