+++ src/apps/ubeswap/positions.e2e.ts
@@ -1,17 +1,13 @@
-import { getConfig } from '../../api/config'
 import { getPositions } from '../../runtime/getPositions'
 import { NetworkId } from '../../types/networkId'
 import hook from './positions'
 
 describe('getPositionDefinitions', () => {
   it('should get the address definitions successfully', async () => {
-    const config = getConfig()
-
     const positions = await getPositions(
       NetworkId['celo-mainnet'],
       '0x2b8441ef13333ffa955c9ea5ab5b3692da95260d',
       ['ubeswap'],
-      config.GET_TOKENS_INFO_URL,
     )
 
     // Uniswap v2 farm definitions
