+++ src/runtime/getPositions.e2e.ts
@@ -1,17 +1,7 @@
 import { getPositions } from './getPositions'
 import { NetworkId } from '../types/networkId'
-import fs from 'fs'
-import yaml from 'js-yaml'
 
 describe('getPositions', () => {
-  let getTokensInfoUrl: string
-  beforeAll(() => {
-    // use the same API to get tokens info as we use in production
-    const envFileContent = yaml.load(
-      fs.readFileSync('src/api/production.yaml', 'utf8'),
-    ) as Record<string, string>
-    getTokensInfoUrl = envFileContent['GET_TOKENS_INFO_URL']
-  })
   it.each([NetworkId['celo-mainnet'], NetworkId['ethereum-mainnet']])(
     'should get the address positions successfully for networkId %s',
     async (networkId) => {
@@ -19,7 +9,6 @@ describe('getPositions', () => {
         networkId,
         '0x2b8441ef13333ffa955c9ea5ab5b3692da95260d',
         [],
-        getTokensInfoUrl,
       )
       // Simple check to make sure we got some definitions
       expect(positions.length).toBeGreaterThan(0)
@@ -36,7 +25,6 @@ describe('getPositions', () => {
         networkId,
         '0x2b8441ef13333ffa955c9ea5ab5b3692da95260d',
         [appId],
-        getTokensInfoUrl,
       )
       // Simple check to make sure we got some definitions
       expect(positions.length).toBeGreaterThan(0)
@@ -52,7 +40,6 @@ describe('getPositions', () => {
         NetworkId['celo-mainnet'],
         '0x2b8441ef13333ffa955c9ea5ab5b3692da95260d',
         ['does-not-exist'],
-        getTokensInfoUrl,
       ),
     ).rejects.toThrow(
       /No app with id 'does-not-exist' found, available apps: \w+/,
@@ -62,7 +49,6 @@ describe('getPositions', () => {
         NetworkId['ethereum-mainnet'],
         '0x2b8441ef13333ffa955c9ea5ab5b3692da95260d',
         ['does-not-exist'],
-        getTokensInfoUrl,
       ),
     ).rejects.toThrow(
       /No app with id 'does-not-exist' found, available apps: \w+/,
